MT.02 - Matrix Sort
Sean invented a game involving a 2n x 2n matrix where each cell of the
matrix contains an integer. He can reverse any of its rows or columns
any number of times. The goal of the game is to maximize the sum of the
elements in the n x n submatrix located in the upper-left quadrant of
the matrix. Given the initial configurations for Q matrices, help Sean
reverse the rows and columns of each matrix in thebest possible way so
that the sum of the elements in the matrix's upper-left quadrant is
maximal.

Ex. Input:
[112, 42, 83, 119]
[56, 125, 56, 49]
[15, 78, 101, 43]
[62, 98, 114, 108]

Ex. Output: 414